Improve package
- Adding RecordFormatterInterface
- Adding RecordValidatorInterface
- Adding CallableValidatorAdapter to reduce BC breakage
- Adding CallableFormatterAdapter to reduce BC breakage
- Removing the Config subnamespace
