+++ src/InsertionException.php
@@ -20,7 +20,8 @@ use RuntimeException;
  *  Thrown when a data is not added to the Csv Document
  *
  * @package League.csv
- * @since  9.0.0
+ * @since   9.0.0
+ * @author  Ignace Nyamagana Butera <nyamsprod@gmail.com>
  *
  */
 class InsertionException extends RuntimeException
